requirements.txt
django
pillow
django-crispy-forms
opencv-python